將entity移除，修改至model/shop
