+++ src/main/java/com/healthmanagement/service/shop/impl/CartItemServiceImpl.java
@@ -4,8 +4,8 @@ import com.healthmanagement.dao.shop.CartItemDAO;
 import com.healthmanagement.dao.shop.ProductDAO;
 import com.healthmanagement.dto.shop.CartItemDTO;
 import com.healthmanagement.dto.shop.CartItemRequest;
-import com.healthmanagement.entity.shop.CartItem;
-import com.healthmanagement.entity.shop.Product;
+import com.healthmanagement.model.shop.CartItem;
+import com.healthmanagement.model.shop.Product;
 import com.healthmanagement.service.shop.CartItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
