+++ src/main/java/com/healthmanagement/service/shop/impl/OrderServiceImpl.java
@@ -6,10 +6,10 @@ import com.healthmanagement.dao.shop.ProductDAO;
 import com.healthmanagement.dto.shop.CreateOrderRequest;
 import com.healthmanagement.dto.shop.OrderDTO;
 import com.healthmanagement.dto.shop.OrderItemDTO;
-import com.healthmanagement.entity.shop.Order;
-import com.healthmanagement.entity.shop.OrderItem;
-import com.healthmanagement.entity.shop.CartItem;
-import com.healthmanagement.entity.shop.Product;
+import com.healthmanagement.model.shop.Order;
+import com.healthmanagement.model.shop.OrderItem;
+import com.healthmanagement.model.shop.CartItem;
+import com.healthmanagement.model.shop.Product;
 import com.healthmanagement.service.shop.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
